+++ modules/home/Home.module.css
@@ -5,10 +5,6 @@
   align-items: center;
 }
 
-.mainTitle::after {
-  content: none;
-}
-
 .secondaryTitle {
   margin-block: 100px 70px;
 }
